VED-386: filevalidation refactoring
